# docs-final/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/search/providers/{providerId}/remarks": {
      "get": {
        "description": "Customer remarks from verified jobs, newest first, shown as first name and last initial only, each with the provider’s reply if they gave one. `reply` is an object `{ body, createdAt }` when there is a reply and `null` when there is not — it is not a string. Unpublished remarks never appear.",
        "operationId": "ReputationController_remarks",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "1–100, default 20",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "A provider’s published remarks",
        "tags": [
          "reputation"
        ]
      }
    },
    "/api/v1/provider/ratings": {
      "get": {
        "description": "Every rating you have received with its four criteria, and the remark that came with it (including any an admin unpublished), your reply, and your overall reputation.",
        "operationId": "ReputationController_own",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "See the ratings I received",
        "tags": [
          "reputation"
        ]
      }
    },
    "/api/v1/provider/remarks/{id}/reply": {
      "post": {
        "description": "One reply per published remark, up to 1000 characters, visible next to it. It cannot be edited afterwards; a second reply is 409.",
        "operationId": "ReputationController_reply",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "body": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 1000
                  }
                },
                "required": [
                  "body"
                ],
                "additionalProperties": false
              },
              "examples": {
                "default": {
                  "summary": "A courteous reply",
                  "value": {
                    "body": "Thank you — glad the leak is fixed."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Reply to a remark",
        "tags": [
          "reputation"
        ]
      }
    },
    "/api/v1/admin/providers/{providerId}/ratings": {
      "get": {
        "description": "Admin only. Every rating and remark for a provider, unpublished ones included with who removed them and why, and their overall reputation.",
        "operationId": "ReputationController_adminRatings",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Review a provider’s ratings",
        "tags": [
          "reputation"
        ]
      }
    },
    "/api/v1/admin/remarks/{id}/unpublish": {
      "post": {
        "description": "Admin only. Hides the remark from the public; the rating behind it stays in the provider’s score. The reason is recorded and the action audited.",
        "operationId": "ReputationController_unpublish",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 500
                  }
                },
                "required": [
                  "reason"
                ],
                "additionalProperties": false
              },
              "examples": {
                "default": {
                  "summary": "Abusive language",
                  "value": {
                    "reason": "Contains personal abuse"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Unpublish an abusive remark",
        "tags": [
          "reputation"
        ]
      }
    },
...
  },
  "info": {
    "title": "Smart Home Maintenance Services API",
    "description": "The backend API for the Smart Home Maintenance Service platform: account sign-up and login, admin settings, payment webhooks, and health checks.\n\n**Getting started:** most endpoints need you to be logged in. Call `POST /auth/login` (password) or the OTP endpoints to get an `accessToken`, then click \"Authorize\" above and enter it as `Bearer <token>` to unlock the endpoints marked with a lock icon.\n\nEndpoints under **development** only work when the API is running with `DEV_INBOX_ENABLED=true` (local/dev setups) — they let you see the text messages, emails and files the mock providers would otherwise send to real services, which is how you retrieve OTP codes while testing.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}